Separacion de exams Student y teacher

+++ app/Http/Controllers/ExamController.php
@@ -25,7 +25,11 @@ class ExamController extends Controller
         $exams=Exam::all();
         return view('exams.index',['exams'=>$exams]);
     }
-
+    public function examsIndexTeacher()
+    {
+        $exams=Exam::all();
+        return view('exams/examsIndexTeacher',['exams'=>$exams]);
+    }
     /**
      * Show the form for creating a new resource.
      *
@@ -37,7 +41,11 @@ class ExamController extends Controller
         return view('exams.create',['patients'=>$patients]);
     }
 
-
+    public function examsCreateTeacher()
+    {
+        $patients = Patient::all()->pluck('name','id');
+        return view('exams/examsCreateTeacher',['patients'=>$patients]);
+    }
     /**
      * Store a newly created resource in storage.
      *
@@ -95,6 +103,45 @@ class ExamController extends Controller
         return redirect()->route('create_asociacionED',[$exam->id]);
     }
 
+    public function examsStoreTeacher(Request $request)
+    {
+        //dd($request);
+        $this->validate($request, [
+            'date'=>['required','date'],
+            'aspectoExtraoralNormal' => ['required', 'boolean'],
+            'cancerOral' => ['required', 'boolean'],
+            'anomaliasLabios' => ['required', 'boolean'],
+            'otros' => ['nullable','string', 'max:255'],
+            'patologiaMucosa'=> ['string','in:Ninguna,Tumor maligno,leucoplasia,Liquen plano'],
+            'fluorosis'=> ['required', 'string','in:Normal,Discutible,Muy ligera,Ligera,Moderada,Intensa,Excluida,No registrada'],
+            'estadoS1'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS2'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS3'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS4'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS5'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS6'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'claseAngle'=> ['required', 'string','in:clase I,clase II,clase III'],
+            'lateralAngle'=> ['required', 'string','in:Unilateral,Bilateral'],
+            'tipoDentición'=> ['required', 'string','in:Temporal,Mixta'],
+            'apiñamientoIncisivoInferior' => ['required', 'boolean'],
+            'apiñamientoIncisivoSuperior' => ['required', 'boolean'],
+            'perdidaEspacioAnterior' => ['required', 'boolean'],
+            'perdidaEspacioPosterior' => ['required', 'boolean'],
+            'mordidaCruzadaAnterior' => ['required', 'boolean'],
+            'mordidaCruzadaPosterior' => ['required', 'boolean'],
+            'desviacionLineaMedia' => ['required', 'boolean'],
+            'mordidaAbierta' => ['required', 'boolean'],
+            'habitos' => ['required', 'boolean'],
+            'patient_id' => ['required', 'exists:patients,id'],
+        ]);
+
+        $exam = new Exam($request->all());
+        $exam->save();
+
+        flash('Examen creado correctamente');
+
+        return redirect()->route('create_asociacionED',[$exam->id]);
+    }
     /**
      * Display the specified resource.
      *
@@ -121,7 +168,13 @@ class ExamController extends Controller
 
         return view('exams/edit',['exam'=> $exam, 'patients'=>$patients ]);
     }
+    public function examsEditTeacher($id)
+    {
+        $exam = Exam::find($id);
+        $patients = Patient::all()->pluck('name','id');
 
+        return view('exams/examsEditTeacher',['exam'=> $exam, 'patients'=>$patients ]);
+    }
     /**
      * Update the specified resource in storage.
      *
@@ -182,6 +235,49 @@ class ExamController extends Controller
 
         return redirect()->route('exams.index');
     }
+    public function examsUpdateTeacher(Request $request, $id)
+    {
+
+        $this->validate($request, [
+            'date'=>['required','date'],
+            'aspectoExtraoralNormal' => ['required', 'boolean'],
+            'cancerOral' => ['required', 'boolean'],
+            'anomaliasLabios' => ['required', 'boolean' ],
+            'otros' => ['nullable','string', 'max:1000'],
+            'patologiaMucosa'=> ['string','in:Ninguna,Tumor maligno,leucoplasia,Liquen plano'],
+            'fluorosis'=> ['required', 'string','in:Normal,Discutible,Muy ligera,Ligera,
+                Moderada,Intensa,Excluida,No registrada'],
+            'estadoS1'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS2'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS3'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS4'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS5'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'estadoS6'=> ['required', 'string','in:sano,hemorragia,tártaro,bolsa 4-5 mm,Bolsa de 6 mm o más,excluido'],
+            'claseAngle'=> ['required', 'string','in:clase I,clase II,clase III'],
+            'lateralAngle'=> ['required', 'string','in:Unilateral,Bilateral'],
+            'tipoDentición'=> ['required', 'string','in:Temporal,Mixta'],
+            'apiñamientoIncisivoInferior' => ['required', 'boolean'],
+            'apiñamientoIncisivoInferior' => ['required', 'boolean'],
+            'apiñamientoIncisivoSuperior' => ['required', 'boolean'],
+            'perdidaEspacioAnterior' => ['required', 'boolean'],
+            'perdidaEspacioPosterior' => ['required', 'boolean'],
+            'mordidaCruzadaAnterior' => ['required', 'boolean'],
+            'mordidaCruzadaPosterior' => ['required', 'boolean'],
+            'desviacionLineaMedia' => ['required', 'boolean'],
+            'mordidaAbierta' => ['required', 'boolean'],
+            'habitos' => ['required', 'boolean'],
+            'patient_id' => ['required', 'exists:patients,id'],
+        ]);
+
+        $exam = Exam::find($id);
+        $exam->fill($request->all());
+
+        $exam->save();
+
+        flash('Examen modificado correctamente');
+
+        return redirect()->route('examsIndexTeacher');
+    }
 
     /**
      * Remove the specified resource from storage.
@@ -189,12 +285,45 @@ class ExamController extends Controller
      * @param  \App\Exam  $exam
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
-            $exam = Exam::find($id);
-            $exam->delete();
-            flash('Examen borrado correctamente');
 
-            return redirect()->route('exams.index');
+
+    public function examsdestroyStudent($id){
+        return view('exams/destroyStudent',['id'=>$id]);
+    }
+    public function examsdeleteStudent(Request $request,$id){
+
+        $this->validate($request, [
+            'pin'=>['required','string','max:255']
+        ]);
+
+        $profesor=DB::select(DB::raw('SELECT * FROM laravel.users
+        LEFT JOIN laravel.asociacion_teacher_students ON (laravel.asociacion_teacher_students.student_id = users.id)
+        LEFT JOIN laravel.users as teachers ON (teachers.id = laravel.asociacion_teacher_students.teacher_id)
+        WHERE laravel.users.id ='.Auth::user()->id.' AND teachers.pin='.$request->get('pin').';'));
+
+        if(count($profesor)==0){
+            flash('Pin incorrecto');
+            return redirect()->route('examsdestroyStudent',['id'=>$id]);
+        }
+
+        $exam = Exam::find($id);
+        $exam->delete();
+        flash('Examen borrado correctamente');
+
+        return redirect()->route('exams.index');
+    }
+
+    public function examsdeleteTeacher($id)
+    {
+        $exam = Exam::find($id);
+        $exam->delete();
+        flash('Examen borrado correctamente');
+
+        return redirect()->route('examsIndexTeacher');
+    }
+
+    public function destroy()
+    {
+
     }
 }
